Eclipse type enum moved in with Eclipse data type
