+++ libecl/include/ert/ecl/ecl_util.h
@@ -27,6 +27,7 @@ extern "C" {
 #include <ert/util/stringlist.h>
 #include <ert/util/time_t_vector.h>
 #include <ert/util/parser.h>
+#include <ert/ecl/ecl_type.h>
 
 
 typedef enum { ECL_OTHER_FILE           = 0   ,
@@ -54,69 +55,6 @@ typedef enum { ECL_OTHER_FILE           = 0   ,
                  ECL_FORMATTED_UNIFIED     = 2,
                  ECL_BINARY_NON_UNIFIED    = 4,
                  ECL_FORMATTED_NON_UNIFIED = 8} ecl_storage_enum;
-
-/*
-  Character data in ECLIPSE files comes as an array of fixed-length
-  string. Each of these strings is 8 characters long. The type name,
-  i.e. 'REAL', 'INTE', ... , come as 4 character strings.
-*/
-
-#define ECL_STRING10_LENGTH 10    // Initial guess that the type C010 type is a10 character string.
-#define ECL_STRING8_LENGTH   8
-#define ECL_TYPE_LENGTH      4
-#define ECL_KW_HEADER_DATA_SIZE   ECL_STRING8_LENGTH + ECL_TYPE_LENGTH + 4
-#define ECL_KW_HEADER_FORTIO_SIZE ECL_KW_HEADER_DATA_SIZE + 8
-
-/*****************************************************************/
-/*
-   Observe that these type identidiers are (ab)used in both the rms and
-   ert/enkf libraries in situations where ECLIPSE is not at all involved.
-
-   Regarding the type ECL_C010_TYPE: In an SMSPEC file generated by IX
-   there was a keyword with header:
-
-      NAMES   316   C010
-
-   i.e. the type identifier was the string 'C010'. Have not found any
-   further documentation of that type, but assuming it is a 10
-   character long string the file loads ok. The ECL_C010_TYPE is
-   implemented on a very loose basis it is not clear if the
-   ten-character-long-string implementation makes sense, or if that
-   just happened to work for the one example we have encountered so
-   far. For this reason the support for this keyword type is very
-   limited:
-
-     - keywords of this type are not propertly internalized; they are
-       just cleanly skipped by ecl_kw/ecl_file implementation when
-       loading a file.
-
-     - it is not possible to instantiate a keyword of this type with
-       ecl_kw_alloc( ).
-
-     - The type is not exported to Python.
-*/
-
-// TODO: move enum to ecl_type
-typedef enum {
-  ECL_CHAR_TYPE   = 0,
-  ECL_FLOAT_TYPE  = 1,
-  ECL_DOUBLE_TYPE = 2,
-  ECL_INT_TYPE    = 3,
-  ECL_BOOL_TYPE   = 4,
-  ECL_MESS_TYPE   = 5,
-  ECL_C010_TYPE   = 6   // See comment immediately above about the limited support of this type.
-} ecl_type_enum;
-
-#define ECL_TYPE_ENUM_DEFS {.value = 0 , .name = "ECL_CHAR_TYPE"}, \
-{.value = 1 , .name = "ECL_FLOAT_TYPE"} ,                          \
-{.value = 2 , .name = "ECL_DOUBLE_TYPE"},                          \
-{.value = 3 , .name = "ECL_INT_TYPE"},                             \
-{.value = 4 , .name = "ECL_BOOL_TYPE"},                            \
-{.value = 5 , .name = "ECL_MESS_TYPE"}
-
-#define ECL_TYPE_ENUM_SIZE 6
-
-#include <ert/ecl/ecl_type.h>
 
 /*
   The libecl library has been built and tested 99.5% with ECLIPSE100
